style: code reformatting, nitpicking, renaming benchmark
